Move Indexed Enums to UInt

+++ core/src/commonMain/kotlin/maryk/core/processors/datastore/writeTypedValueToStorage.kt
@@ -28,7 +28,7 @@ fun <T : IsPropertyDefinition<*>> writeTypedValueToStorage(
         val qualifierTypeLength = qualifierSize + typedValue.type.index.calculateVarIntWithExtraInfoByteSize()
         val qualifierTypeWriter = createQualifierWriter(
             qualifierWriter,
-            typedValue.type.index,
+            typedValue.type.index.toInt(),
             TYPE
         )
